优化启动逻辑

+++ docker/start.sh
@@ -8,14 +8,12 @@ function setting {
     echo $TZ > /etc/timezone
 
     # 兼容旧config文件路径
-    if [ -d /opt/config ]; then
+    if [ -d /opt/config ] && [ ! -d /config ]; then
         echo -e "使用v1.x版本config路径配置"
         rm -rf /config
         ln -s /opt/config /
     else
-        if [ ! -d /config ]; then
-            mkdir -p /config
-        fi
+        mkdir -p /config
     fi
 }
 if [ ! -f /setting.lock ]; then
@@ -23,6 +21,25 @@ if [ ! -f /setting.lock ]; then
 fi
 
 # 自动更新
+function app_update {
+    echo "更新程序..."
+    git remote set-url origin ${REPO_URL} &>/dev/null
+    git clean -dffx
+    git reset --hard HEAD
+    git pull
+}
+
+function requirement_update {
+    echo "检测到requirement.txt有变化，重新安装依赖..."
+    pip install --upgrade pip setuptools wheel
+    pip install -r requirement.txt
+}
+
+function package_list_update {
+    echo "检测到package_list.txt有变化，更新软件包..."
+    apk add --no-cache $(echo $(cat docker/package_list.txt))
+}
+
 if [ "${EmbyTools_AUTO_UPDATE}" = "true" ]; then
     if [ ! -s /tmp/requirement.txt.sha256sum ]; then
         sha256sum requirement.txt > /tmp/requirement.txt.sha256sum
@@ -30,19 +47,13 @@ if [ "${EmbyTools_AUTO_UPDATE}" = "true" ]; then
     if [ ! -s /tmp/package_list.txt.sha256sum ]; then
         sha256sum docker/package_list.txt > /tmp/package_list.txt.sha256sum
     fi
-    echo "更新程序..."
-    git remote set-url origin ${REPO_URL} &>/dev/null
-    git clean -dffx
-    git reset --hard HEAD
-    git pull
+    app_update
     if [ $? -eq 0 ]; then
         echo "更新成功..."
         hash_old=$(cat /tmp/requirement.txt.sha256sum)
         hash_new=$(sha256sum requirement.txt)
         if [ "$hash_old" != "$hash_new" ]; then
-            echo "检测到requirement.txt有变化，重新安装依赖..."
-            pip install --upgrade pip setuptools wheel
-            pip install -r requirement.txt
+            requirement_update
             if [ $? -ne 0 ]; then
                 echo "无法安装依赖，请更新镜像..."
             else
@@ -51,8 +62,7 @@ if [ "${EmbyTools_AUTO_UPDATE}" = "true" ]; then
                 hash_old=$(cat /tmp/package_list.txt.sha256sum)
                 hash_new=$(sha256sum docker/package_list.txt)
                 if [ "$hash_old" != "$hash_new" ]; then
-                    echo "检测到package_list.txt有变化，更新软件包..."
-                    apk add --no-cache $(echo $(cat docker/package_list.txt))
+                    package_list_update
                     if [ $? -ne 0 ]; then
                         echo "无法更新软件包，请更新镜像..."
                     else
@@ -70,8 +80,16 @@ else
 fi
 
 # 权限设置
-chown -R ${PUID}:${PGID} ${WORK_DIR} ${EMBYTOOLS_CONFIG}
-chmod +x ${WORK_DIR}/docker/start.sh
+chown -R ${PUID}:${PGID} ${WORK_DIR}
+
+if [[ "$(stat -c '%u' ${EMBYTOOLS_CONFIG})" != "${PUID}" ]] || [[ "$(stat -c '%g' ${EMBYTOOLS_CONFIG})" != "${PGID}" ]]; then
+    chown ${PUID}:${PGID} ${EMBYTOOLS_CONFIG}
+fi
+
+if [[ "$(stat -c '%A' ${WORK_DIR}/docker/start.sh)" != "-rwxr-xr-x" ]]; then
+chmod 755 ${WORK_DIR}/docker/start.sh
+fi
+
 umask ${UMASK}
 
 # 启动
